feat: move api out of package
